t1404: don't create unused file
Some tests in file t1404-update-ref-errors.sh create file "unchanged" as
the expected side for a test_cmp assertion at the end of the test for
output of "git for-each-ref".  The filename conveys the expectation that
the output won't change between two invocations of "git for-each-ref".

Test 'no bogus intermediate values during delete' also creates a file
named "unchanged".  However, in this test the reference is being
deleted, i.e. it _does change_.  The file itself isn't used for any
assertions in the test.

Don't create the unused and slightly misleading file "unchanged".

Signed-off-by: Andrei Rybak <rybak.a.v@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>