rebase: add a config option for --rebase-merges
The purpose of the new option is to accommodate users who would like
--rebase-merges to be on by default and to facilitate turning on
--rebase-merges by default without configuration in a future version of
Git.

Name the new option rebase.rebaseMerges, even though it is a little
redundant, for consistency with the name of the command line option and
to be clear when scrolling through values in the [rebase] section of
.gitconfig.

In the future, the default rebase-merges mode may change from
no-rebase-cousins to some other mode that doesn't exist yet. Support
setting rebase.rebaseMerges to the nonspecific value "true" for users
who do not need or want to care about the default changing in the
future. Similarly, for users who have --rebase-merges in an alias and
want to get the future behavior now, use the specific rebase-merges mode
from the config if a specific mode is not given on the command line.

Signed-off-by: Alex Henrie <alexhenrie24@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>